Operate on batches of shots
Rather than processing shots individually, now the propagators (and rest
of the code) are designed to work with batches of shots. A batch can
have a single shot, if the previous behaviour is desired.

This is mainly motivated by the expected future implementation of GPU
propagators which are likely to perform better on batches of shots when
the computational domain of each shot is small.

This commit also fixes a bug in Scalar2D (the PML indexing was off),
removes Scalar1D_16, and separates the tests in models_scalar into a new
file test_models_scalar so that pytest can find them.